+++ CHANGELOG.md
@@ -6,7 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2021-06-29
+## [0.1.1] - 2021-08-21
+## Fixed
+- Fixed the config file save failed error
+
+## [0.1.0] - 2021-08-20
+### Added
+- Finished event path view
+- Added save and load function
+- Added directory expand and fold view
+
+## [0.0.1] - 2021-08-19
 ### Added
 - Initialize repository
 - Finish implement basic invoke and listen function